These all have what's necessary for my new resolution-control architecture; screen.h will need more size definitions, and screen.c will need more table entries and some appropriate screen portRect adjustments.
Got the HUD properly placed
Got the resizing worked out; it works successfully in software and OpenGL modes; with this exception:

Piping 2D stuff through OpenGL is still somewhat buggy.

Added support for telling Pfhortran what's the corresponding resolution with or without a HUD

Did correct placement of HUD in all supported resolution; also changed the writing of all the 2D stuff in OpenGL mode -- it's now written to the front buffer. This make help in rendering for 3dfx cards.


git-svn-id: svn://svn.code.sf.net/p/marathon/code/trunk@69 a75d40c9-0969-4759-85e5-97cb871b7ed4