Abstract base class for handling various tasks associated with files
Modified it to be a concrete one, and one with two classes (FileObject and OpenedFile) -- and at Christian Bauer's suggestion, put the MacOS-specific stuff in it, to be selected with preprocessor statements.

Added classes for opened resource files, loaded resource handles, etc.
Got SDL support in it

Added a method for getting the system-default color table to LoadedResource

Adjusted for CB's changes to the overloading of "=" for FileSpecifier


git-svn-id: svn://svn.code.sf.net/p/marathon/code/trunk@312 a75d40c9-0969-4759-85e5-97cb871b7ed4