Added annotation that the original author of most of these is unknown, but is probably Bo Lindbergh.
Added include guards to all the header files.

Placed CB's Mac type declarations inside of cstypes.h, and placed that in front of some other definitions.

Has some template functions for doing memcpy and memset on objects and sets of objects

Modified the obj_copy functions so that "source" can be a constant

Added ()'s to TEST_FLAG implementation, a standard C-preprocessor precaution.


git-svn-id: svn://svn.code.sf.net/p/marathon/code/branches/LP_fileio@346 a75d40c9-0969-4759-85e5-97cb871b7ed4