Overhaul DEBUG_FAST_CODE
1) Removes ad hoc redefinitions of [v]assert and replaces usage with
fc_[v]assert, defined just once in csalerts.h, to cut down on
boilerplate for current and future code and to allow for clean usage
within headers.

Side effect: fixes bug in world.cpp where all the assert calls are
enabled regardless of DEBUG_FAST_CODE because cseries.h is included
after, not before, the ad hoc assert definition.

Note: [v]warn, which is defined alongside [v]assert, is not given an fc_
version because it should never be called with high frequency anyway, to
avoid potentially spamming stderr.

2) Removes conditional "#undef DEBUG" lines to be logically consistent
with normal assert calls, if any need to be added in the future, and to
avoid potentially violating C++'s One Definition Rule if any included
headers later add inline functions with external linkage that depend on
DEBUG. No immediate impact.

3) Removes the conditional macro definition for csprintf (hiding the
function of the same name) in scottish_textures.cpp to cleanly allow
normal csprintf function calls to be added in the future, if needed. No
immediate impact.

4) Renames the conditional vhalt macro (hiding the function of the same
name) in scottish_textures.cpp to VHALT_DEBUG to clarify intent and to
cleanly allow normal vhalt function calls to be added in the future, if
needed. Also adjusts the #if condition to be compiler-agnostic. No
immediate impact (under GCC).