Fix VorbisDecoder segfault and resource leaks
For class VorbisDecoder:

(1) Zero-initializes .ov_file in the ctor so that .Close() won't
segfault if called before a successful call to .Open().

(2) Refactors .Open() so that ov_file and the RWops object always get
cleaned up on failure or if .Open() is called twice in succession.

(3) Simplifies the calls to ov_test_callbacks() and ov_test_open() to a
single call to ov_open_callbacks().