Set local/current player immediately
Set the local and current player indexes/pointers to NONE/nullptr
before player creation and assign them immediately on creating the
local/current players, allowing player creation code to make use of
that information.

Add a general flags parameter to new_player() and an initial set of
flags for setting the local and/or current player.