Restore original mouse pitch acceleration
When the input_flags scaling asymmetry was removed, the sensitivity in old prefs was divided by 4 to compensate. The new sensitivity results in different values post-acceleration, since sensitivity is part of the acceleration calculation. This commit adjusts the vertical acceleration to 1.2-style levels - at the new sens of 0.25, the same post-acceleration values are seen as in 1.2 with a sens of 1.0. It also adds another acceleration type so 1:1 horizontal and vertical behavior can be achieved with acceleration.