Eliminate static from inlined alert functions
I have encountered several cases at work where customers have used
static in headers which causes each translation unit to get a copy of
the given function body. It wastes space and is totally useless in
modern C++ (c++11 and greater). Just marking a function as inline is
enough these days to get the compiler to inline it. Even if the compiler
decides to not inline the function then it can make one instance of the
body and stash it somewhere (although I haven't seen this on gcc or
clang based compilers).