Give all clipping windows valid y clips
Due solely to rounding errors, the x range of a clipping window will
sometimes not overlap the x range of the polygon it is for (in integral
classic screen coordinates). This patch fixes not initializing the y
clips of such clipping windows, which also fixes undefined behavior
downstream (uninitialized reads and integer overflows) and very rare
polygon rendering glitches by the software renderer. OpenGL rendering
is entirely unaffected, since it ignores y clips.